requirements.txt
apache-sedona==1.5.1
asttokens==2.4.1
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
confluent-kafka==2.3.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastavro==1.9.3
idna==3.6
ipython==8.18.1
jedi==0.19.1
kafka-python==2.0.2
matplotlib-inline==0.1.6
numpy==1.26.3
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
Pygments==2.17.2
pyspark==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
shapely==2.0.2
six==1.16.0
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
